# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: dongled
  version: 1.0.0
  description: >-
    Operations panel and customer API for a farm of Huawei E3372 HiLink dongles sold as
    SOCKS5/HTTP proxies. Customer-facing identifiers are always proxy_id, never dongle_id.
servers:
  - url: /
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "204":
          description: session cookie issued
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/auth/session:
    get:
      tags: [auth]
      operationId: session
      responses:
        "200":
          description: current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/proxies:
    get:
      tags: [proxies]
      operationId: listProxies
      parameters:
        - name: customer_id
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProxyState"
        - name: expiring_within_days
          in: query
          required: false
          description: >-
            Keeps proxies whose expires_at is at or before now plus this many days. Already expired
            proxies are included, because they are the ones an operator most needs to see.
          schema:
            type: integer
      responses:
        "200":
          description: proxy list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyList"
  /api/v1/proxies/export:
    get:
      tags: [proxies]
      operationId: exportProxies
      description: >-
        txt emits host:port:user:pass, one proxy per line, with no trailing newline. csv emits the
        header row host,port,username,password followed by the same fields, RFC4180 quoted. A proxy
        whose auth_mode is iplist carries no credentials and is emitted as host:port in txt and with
        two empty trailing fields in csv. A userpass or both proxy whose password is missing is left
        out entirely; the count of emitted and skipped rows is returned in X-Export-Rows and
        X-Export-Skipped.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [txt, csv]
        - name: scheme
          in: query
          required: false
          schema:
            type: string
            enum: [socks5, http]
        - name: ids
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: one proxy per line as host:port:user:pass
          headers:
            Content-Disposition:
              schema:
                type: string
            X-Export-Rows:
              schema:
                type: integer
            X-Export-Skipped:
              schema:
                type: integer
          content:
            text/plain:
              schema:
                type: string
            text/csv:
              schema:
                type: string
  /api/v1/proxies/{proxy_id}:
    get:
      tags: [proxies]
      operationId: getProxy
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      responses:
        "200":
          description: proxy detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/proxies/{proxy_id}/rotate:
    post:
      tags: [proxies]
      operationId: rotateProxyAdmin
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      responses:
        "202":
          description: rotation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "409":
          $ref: "#/components/responses/OpInProgress"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/proxies/{proxy_id}/auth:
    post:
      tags: [proxies]
      operationId: setProxyAuth
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAuthRequest"
      responses:
        "202":
          description: auth change accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/OpInProgress"
  /api/v1/proxies/{proxy_id}/auth-ips:
    get:
      tags: [proxies]
      operationId: listProxyAuthIPs
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      responses:
        "200":
          description: whitelisted source networks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthIPList"
    post:
      tags: [proxies]
      operationId: addProxyAuthIP
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthIPRequest"
      responses:
        "200":
          description: updated whitelist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthIPList"
        "400":
          $ref: "#/components/responses/Error"
    delete:
      tags: [proxies]
      operationId: deleteProxyAuthIP
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthIPRequest"
      responses:
        "200":
          description: updated whitelist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthIPList"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/proxies/{proxy_id}/ports:
    post:
      tags: [proxies]
      operationId: setProxyPorts
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyPolicy"
      responses:
        "202":
          description: port policy accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/OpInProgress"
  /api/v1/proxies/{proxy_id}/enable:
    post:
      tags: [proxies]
      operationId: setProxyEnabled
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableRequest"
      responses:
        "200":
          description: proxy patched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proxy"
  /api/v1/proxies/{proxy_id}/customer:
    post:
      tags: [proxies]
      operationId: assignProxyCustomer
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignCustomerRequest"
      responses:
        "200":
          description: proxy patched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proxy"
  /api/v1/proxies/{proxy_id}/selftest:
    post:
      tags: [proxies]
      operationId: selftestProxy
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      responses:
        "200":
          description: selftest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelftestResult"
...
  /api/v1/dongles/{dongle_id}:
    get:
      tags: [dongles]
      operationId: getDongle
      parameters:
        - $ref: "#/components/parameters/DongleID"
      responses:
        "200":
          description: dongle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DongleDetail"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [dongles]
      operationId: patchDongle
      parameters:
        - $ref: "#/components/parameters/DongleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonglePatchRequest"
      responses:
        "200":
          description: dongle patched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dongle"
  /api/v1/dongles/{dongle_id}/reboot:
    post:
      tags: [dongles]
      operationId: rebootDongle
      parameters:
        - $ref: "#/components/parameters/DongleID"
      responses:
        "202":
          description: reboot accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "409":
          $ref: "#/components/responses/OpInProgress"
  /api/v1/dongles/{dongle_id}/netmode:
    post:
      tags: [dongles]
      operationId: setDongleNetMode
      parameters:
        - $ref: "#/components/parameters/DongleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NetModeRequest"
      responses:
        "202":
          description: net mode change accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "409":
          $ref: "#/components/responses/OpInProgress"
  /api/v1/dongles/{dongle_id}/lanip:
    post:
      tags: [dongles]
      operationId: setDongleLanIP
      parameters:
        - $ref: "#/components/parameters/DongleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanIPRequest"
      responses:
        "202":
          description: lan ip change accepted, expect a re_discovering step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "409":
          $ref: "#/components/responses/OpInProgress"
  /api/v1/dongles/{dongle_id}/sms:
    get:
      tags: [dongles]
      operationId: listSms
      parameters:
        - $ref: "#/components/parameters/DongleID"
        - name: box
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SmsBox"
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: size
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: sms page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsList"
  /api/v1/dongles/{dongle_id}/sms/send:
    post:
      tags: [dongles]
      operationId: sendSms
      parameters:
        - $ref: "#/components/parameters/DongleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsSendRequest"
      responses:
        "204":
          description: queued on the modem
  /api/v1/dongles/{dongle_id}/sms/delete:
    post:
      tags: [dongles]
      operationId: deleteSms
      parameters:
        - $ref: "#/components/parameters/DongleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsIndexRequest"
      responses:
        "204":
          description: deleted
  /api/v1/dongles/{dongle_id}/sms/read:
    post:
      tags: [dongles]
      operationId: markSmsRead
      parameters:
        - $ref: "#/components/parameters/DongleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsIndexRequest"
      responses:
        "204":
          description: marked read
  /api/v1/operations:
    get:
      tags: [operations]
      operationId: listOperations
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OpKind"
        - name: trigger
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Trigger"
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OpState"
        - name: subject_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: operation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationList"
  /api/v1/operations/{op_id}:
    get:
      tags: [operations]
      operationId: getOperation
      parameters:
        - name: op_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/rotations:
    get:
      tags: [operations]
      operationId: listRotations
      parameters:
        - name: proxy_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: redacted rotation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotationList"
  /api/v1/customers:
    get:
      tags: [customers]
      operationId: listCustomers
      responses:
        "200":
          description: customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      tags: [customers]
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /api/v1/customers/{customer_id}:
    patch:
      tags: [customers]
      operationId: patchCustomer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        "200":
          description: patched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /api/v1/keys:
    get:
      tags: [keys]
      operationId: listApiKeys
      responses:
        "200":
          description: api key list, hashes never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyList"
    post:
      tags: [keys]
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyRequest"
      responses:
        "201":
          description: created, secret shown exactly once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
  /api/v1/keys/{key_id}:
    delete:
      tags: [keys]
      operationId: revokeApiKey
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: revoked
  /api/v1/keys/{key_id}/link-tokens:
    post:
      tags: [keys]
      operationId: createLinkToken
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: created, url shown exactly once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkTokenCreated"
  /api/v1/link-tokens/{token_id}:
    delete:
      tags: [keys]
      operationId: revokeLinkToken
      parameters:
        - name: token_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: revoked independently of the api key
...
  /api/v1/rotate/{proxy_id}:
    post:
      tags: [customer-api]
      operationId: rotateProxyCustomer
      description: >-
        Asynchronous by default: 202 with an operation id and a poll url. wait=true blocks
        for at most 90s and returns the terminal result. A rotation that completes without
        changing the public IP is result unchanged with ip_changed false.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProxyID"
        - name: wait
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: terminal result when wait=true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateResult"
        "202":
          description: accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/OpInProgress"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/status/{proxy_id}:
    get:
      tags: [customer-api]
      operationId: customerStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProxyID"
      responses:
        "200":
          description: customer visible status, never exposes dongle_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerStatus"
        "401":
          $ref: "#/components/responses/Error"
  /r/{link_token}:
    get:
      tags: [customer-api]
      operationId: linkConfirmPage
      description: >-
        Renders a confirmation page. GET has no side effect because link scanners and
        prefetchers do not send any identifying header.
      parameters:
        - $ref: "#/components/parameters/LinkToken"
      responses:
        "200":
          description: confirmation page
          content:
            text/html:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [customer-api]
      operationId: linkRotate
      parameters:
        - $ref: "#/components/parameters/LinkToken"
      responses:
        "200":
          description: terminal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateResult"
        "202":
          description: accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "409":
          $ref: "#/components/responses/OpInProgress"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-dongled_session
      description: >-
        SameSite=Strict, HttpOnly, Secure. Every request that is not GET, HEAD or OPTIONS must
        also carry the header X-CSRF-Token holding Session.csrf_token. POST /auth/login is exempt
        because no session exists yet, and POST /auth/logout is exempt because a forged sign out
        is not worth breaking a client that has not yet read its session.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    ProxyID:
      name: proxy_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    OpInProgress:
      description: another operation already runs on this subject
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpInProgress"
    RateLimited:
      description: >-
        rate limited. Retry-After and Error.retry_after always carry the same number of seconds,
        so a client may read either. Per key limits also set the X-RateLimit headers; the per proxy
        minimum rotation interval does not, because it is a product policy rather than a quota.
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    Session:
      type: object
      required: [username, expires_at]
      properties:
        username:
          type: string
        expires_at:
          type: integer
        csrf_token:
          type: string
          description: send this verbatim as X-CSRF-Token on every request that is not GET, HEAD or OPTIONS
...
    ProxyPolicy:
      type: object
      required: [allow_all_ports, max_conn, conn_limit]
      properties:
        allow_all_ports:
          type: boolean
        allowed_ports:
          type: array
          items:
            $ref: "#/components/schemas/PortRange"
        max_conn:
          type: integer
        conn_limit:
          type: integer
...
    ProxyState:
      type: string
      enum: [active, suspended, disabled, expired, degraded, unknown]
...
    SmsBox:
      type: integer
      enum: [1, 2, 3]
...
    AuthIPList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuthIP"
    AuthIPRequest:
      type: object
      required: [cidr]
      properties:
        cidr:
          type: string
        note:
          type: string
...
    ProxyList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Proxy"
        total:
          type: integer
    ProxyDetail:
      type: object
      required: [proxy]
      properties:
        proxy:
          $ref: "#/components/schemas/Proxy"
        auth_ips:
          type: array
          items:
            $ref: "#/components/schemas/AuthIP"
        slot:
          $ref: "#/components/schemas/Slot"
        last_rotation:
          $ref: "#/components/schemas/Rotation"
    SetAuthRequest:
      type: object
      required: [auth_mode]
      properties:
        auth_mode:
          $ref: "#/components/schemas/AuthMode"
        username:
          type: string
        password:
          type: string
        rotate_password:
          type: boolean
    EnableRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    AssignCustomerRequest:
      type: object
      properties:
        customer_id:
          type: string
          nullable: true
        expires_at:
          type: integer
          nullable: true
...
    OperationAccepted:
      type: object
      required: [operation_id, poll_url]
      properties:
        operation_id:
          type: string
        poll_url:
          type: string
        state:
          $ref: "#/components/schemas/OpState"
        deadline_at:
          type: integer
...